Javadoc improved + IdentifyGitOriginRepoTask moved to a more proper package #302
